Generate.py  ( Gui interface for Imaginepy )
This is a GUI interface for Imaginepy.  I took the app.py and re write it with a FULL blown GUI. I used Tkinter for the gui to make sure maximum support for both linux and windows.

it has user input box, selection boxes for model, style and ratio.
it has an image preview that shows on the interface when image is generated. if the preview is clicked it opens in default image viewer. ( both windows and linux supported ).
it has a browse button that lets you browse the images folder ( both windows and linux supported ).
it has a pop up "generating image..." window with aniation to show the image is being generated.

its a pretty nice GUI, fully operational. ( ill keep working on it to add more features )